WebGLRenderer: move conversion methods to utils
